Added documentation and refactoring notes to Class.cs and Method.cs.
After looking into cecil, they made a design decision to just extend MethodSpecification by adding the class GenericInstanceMethod and having it inherit MethodSpecification.  They did this to prevent any extra fields from being exposed by MethodSpecification that would not be used by non-generic methods.  But at the same time, this design decision prevents a single class from representing all of the method information present. As a side effect they are causing branching statements in projects (like SharpOS) that has to deal with both situations, which is thus increasing the code complexity and making this code base less stable.  

In the future, a tool needs to be created that combines the information from both Mono.Cecil.MethodSpecification and Mono.Cecil.GenericInstanceMethod, and then using that in this code base.  This would help prevent branching statements from arising in the future.